+++ src/code_actions/remove_unused_import.rs
@@ -341,6 +341,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -379,6 +380,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -424,6 +426,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -459,6 +462,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -495,6 +499,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -552,6 +557,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -644,6 +650,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),
@@ -756,6 +763,7 @@ mod tests {
             context: CodeActionContext {
                 diagnostics: vec![],
                 only: None,
+                trigger_kind: None,
             },
             work_done_progress_params: Default::default(),
             partial_result_params: Default::default(),