+++ src/main.rs
@@ -58,9 +58,7 @@ async fn main() {
     let stdin = tokio::io::stdin();
     let stdout = tokio::io::stdout();
 
-    let (service, socket) = LspService::build(Backend::new)
-        .custom_method("textDocument/inlayHint", Backend::inlay_hint_request)
-        .finish();
+    let (service, socket) = LspService::build(Backend::new).finish();
 
     Server::new(stdin, stdout, socket).serve(service).await;
 }